Move to trait.
